Fix fastbitmap iterators
Created TPixelRect which behaves a little more in line with what a sane person would expect TRect to behave like in the siturations where it points to a rectangle in a pixel grid.  TPixelRect is now used for TFastBitmap's iterators, so there should be no confusion as to the ranges passed to your multi-threaded effects functions.

Enabled TGraphic button to auto-generate "Over" and "Down" images if they are missing using TFastBitmap's iterators.